Moving the classes to the right packages
