+++ app/src/main/java/pl/llp/aircasting/ui/view/fragments/LetsBeginFragment.kt
@@ -1,4 +1,4 @@
-package pl.llp.aircasting.ui.view.screens.lets_begin
+package pl.llp.aircasting.ui.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import pl.llp.aircasting.AircastingApplication
 import pl.llp.aircasting.ui.view.common.BaseFragment
+import pl.llp.aircasting.ui.view.screens.lets_begin.LetsBeginController
+import pl.llp.aircasting.ui.view.screens.lets_begin.LetsBeginViewMvcImpl
 import pl.llp.aircasting.util.exceptions.ErrorHandler
 import pl.llp.aircasting.util.helpers.permissions.PermissionsManager
 import javax.inject.Inject
