+++ app/src/main/java/pl/llp/aircasting/ui/view/fragments/OnboardingHowsTheAirFragment.kt
@@ -1,10 +1,13 @@
-package pl.llp.aircasting.ui.view.screens.onboarding.how_is_the_air
+package pl.llp.aircasting.ui.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import pl.llp.aircasting.ui.view.common.BaseFragment
+import pl.llp.aircasting.ui.view.screens.onboarding.how_is_the_air.OnboardingHowsTheAirController
+import pl.llp.aircasting.ui.view.screens.onboarding.how_is_the_air.OnboardingHowsTheAirViewMvc
+import pl.llp.aircasting.ui.view.screens.onboarding.how_is_the_air.OnboardingHowsTheAirViewMvcImpl
 
 class OnboardingHowsTheAirFragment:  BaseFragment<OnboardingHowsTheAirViewMvcImpl, OnboardingHowsTheAirController>() {
     lateinit var listener: OnboardingHowsTheAirViewMvc.Listener
