+++ app/src/main/java/pl/llp/aircasting/ui/view/fragments/OnboardingMeasureAndMapFragment.kt
@@ -1,10 +1,13 @@
-package pl.llp.aircasting.ui.view.screens.onboarding.measure_and_map
+package pl.llp.aircasting.ui.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import pl.llp.aircasting.ui.view.common.BaseFragment
+import pl.llp.aircasting.ui.view.screens.onboarding.measure_and_map.OnboardingMeasureAndMapController
+import pl.llp.aircasting.ui.view.screens.onboarding.measure_and_map.OnboardingMeasureAndMapViewMvc
+import pl.llp.aircasting.ui.view.screens.onboarding.measure_and_map.OnboardingMeasureAndMapViewMvcImpl
 
 class OnboardingMeasureAndMapFragment: BaseFragment<OnboardingMeasureAndMapViewMvcImpl, OnboardingMeasureAndMapController>() {
     lateinit var listener: OnboardingMeasureAndMapViewMvc.Listener
